\name{immune}
\docType(data}
\alias{immune}

\title{Gene Expression Data for a mouse immune response study}

\usage{
  data(immune)
}

\description{
This data set consists of gene expression for mouse T-cell immune responses (Affymetrix MG-U74vA GeneChip).

}

\format{
 A matrix of 12,488 genes by 10 columns (1=probe set ID, 2,3,4=Naive, 5,6,7=48h activated, and 8,9,10=CD8+ T-cell clone D4).
}


\references{
	Lee, JK and Connell, MO (2003). \emph{An S-Plus library for the analysis of differential expression}. In The Analysis of Gene Expression Data: Methods and Software. Edited by G. Parmigiani, ES Garrett, RA Irizarry ad SL Zegar. Springer, New York.
	
	Jain et. al. (2003) \emph{Local pooled error test for identifying
      differentially expressed genes with a small number of replicated microarrays}, 19: 1945-1951, Bioinformatics. 

 See "http://hsc.virginia.edu/medicine/basic-sci/biomed/ley/"
 or contact Dr. Klaus Ley (kfl3f@virginia.edu)

}

\keyword{datasets}
