\name{plotsample}
\alias{plotsample}
\title{
Plotting MCMC samples for diagnosis
}
\description{
Plots MCMC samples for diagnosis
}

\usage{
function (a)
}

\arguments{
\item{a}{HEM object}
}

\author{
HyungJun Cho and Jae K. Lee
}
\seealso{
\code{\link{hem}}, 
\code{\link{hem.eb.prior}}, 
\code{\link{hem.fdr}} 
}

\keyword{models}
