\name{radiate}
\alias{radiate}
\title{
Gene expression data for ionizing radiation data
}
\description{
This data set consists of gene expression of ionizing radiation response study (Affymetrix HUGENEFL GeneChip). 
}
\usage{data(radiate)}

\format{A matrix containing 7129 probe sets and 8 chips (U1A, U1B, U2A, U2B, I1A, I1B, I2A, I2B). U=Unirradiated state, I=Irradidated state; cell lines 1, 2; aliquots A, B for independent hybridizations}
\source{
Tusher, V.G., Tibshirani, R., and Chu, G. (2001) Significance analysis of microarrays applied to the ionizing radiation response, PNAS 98:5116-5121

See \url{http://www-stat.stanford.edu/~tibs/SAM/}

}
\keyword{datasets}

